requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
alembic==1.0.5
amqp==2.5.2
aniso8601==8.0.0
anyjson==0.3.3
apipkg==1.5
appdirs==1.4.4
arrow==0.10.0
asn1crypto==0.23.0
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
Authlib==0.10
avro-python3==1.9.2.1
Babel==2.4.0
billiard==3.6.3.0
blinker==1.4
cachetools==3.1.0
celery==4.4.2
certifi==2017.7.27.1
cffi==1.9.1
chardet==3.0.4
chargebee==2.3.0
click==6.7
closeio==1.1
colorclass==2.2.0
colorlog==4.1.0
confluent-kafka==1.4.2
contextlib2==0.5.4
createsend==6.1.0
croniter==0.3.32
crypto==1.4.1
cryptography==2.4.2
cssselect==1.0.1
cssutils==1.0.1
datadog==0.38.0
decorator==4.0.11
dnspython==1.16.0
dparse==0.4.1
email-validator==1.1.1
-e git+https://github.com/Nibblecomm/EmmaPython@58bec76f28a8cfc5861fc159b2e765deb33853ca#egg=emma
enum34==1.1.6
execnet==1.5.0
-e git+https://github.com/unifispot/facebook-sdk@b06ee1b4b80e371dc808ac81cfb1633af80a5345#egg=facebook_sdk
factory-boy==2.8.1
Faker==1.0.0
fastavro==0.23.4
faust==1.10.4
Flask==1.1.2
flask-accepts==0.17.2
Flask-Assets==0.12
Flask-Babel==0.11.1
Flask-BabelEx==0.9.4
-e git+https://github.com/sh4nks/flask-babelplus.git@d636fed64d0a722508ebc18c049d1bf5f093ec07#egg=Flask_BabelPlus
Flask-Caching==1.8.0
Flask-Classful==0.14.2
Flask-DebugToolbar==0.10.1
Flask-JWT-Extended==3.20.0
Flask-KVSession==0.6.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Menu==0.7.1
Flask-Principal==0.4.0
Flask-QRcode==0.10.0
Flask-Redis==0.3.0
flask-restx==0.2.0
Flask-Script==2.0.5
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-Themes2==0.1.4
Flask-WTF==0.14.3
flaskerize==0.14.0
fs==2.4.11
funcsigs==1.0.2
future==0.17.1
geoip2==2.4.2
gevent==20.5.0
google-api-core==1.16.0
google-api-python-client==1.6.5
google-auth==1.10.1
google-cloud-core==1.2.0
google-cloud-datastore==1.10.0
google-cloud-logging==1.14.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graypy==2.1.0
greenlet==0.4.15
grpcio==1.20.1
httplib2==0.10.3
idna==2.6
importlib-metadata==1.6.0
inflection==0.3.1
ipaddress==1.0.18
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kafka-python==1.4.7
kombu==4.6.8
linkedin==0.1.5
lxml==4.5.1
mailjet-rest==1.3.0
Mako==1.0.6
MarkupSafe==1.1.1
marshmallow==3.6.0
maxminddb==1.2.3
messagebird==1.2.1
mixpanel==4.4.0
mock==2.0.0
mode==4.3.2
more-itertools==4.3.0
multidict==4.7.6
mypy-extensions==0.4.3
mysqlclient==1.4.6
Naked==0.1.31
natsort==7.0.1
ndg-httpsclient==0.4.2
netaddr==0.7.19
netifaces==0.10.7
neverbounce-sdk==4.2.6
nexmo==1.5.0
numpy==1.16.6
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.1.0
olefile==0.44
opentracing==1.3.0
packaging==16.8
pandas==0.24.2
passlib==1.7.1
pathlib2==2.3.3
paypalrestsdk==1.13.1
pbr==3.0.0
phonenumbers==8.3.2
Pillow==7.1.2
pluggy==0.13.1
postmarker==0.13.0
premailer==3.6.1
protobuf==3.7.1
py==1.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
-e git+https://github.com/unifispot/pycard@7847297287d847ec4f21ef968e9c654017f9d532#egg=pycard
pycountry==17.9.23
pycparser==2.17
pycryptodome==3.8.1
PyJWT==1.5.2
pymongo==3.4.0
pymsteams==0.1.12
pyOpenSSL==18.0.0
pyparsing==2.1.10
pyrad==2.1
pyrsistent==0.16.0
PySocks==1.6.8
pytest==5.4.2
pytest-base-url==1.4.1
pytest-flask==1.0.0
pytest-flask-sqlalchemy==1.0.0
pytest-forked==0.2
pytest-html==1.22.0
pytest-metadata==1.8.0
pytest-mock==1.10.0
pytest-selenium==1.17.0
pytest-variables==1.7.1
pytest-xdist==1.24.1
python-dateutil==2.6.0
python-editor==1.0.3
python-intercom==3.1.0
python-ldap==3.1.0
python-linkedin==4.1
python-linkedin-v2==0.9.4
pytz==2017.2
PyYAML==3.12
qrcode==5.3
raven==6.0.0
redis==3.5.2
requests==2.22.0
requests-cache==0.4.13
requests-oauthlib==1.0.0
robinhood-aiokafka==1.1.6
rsa==3.4.2
safety==1.8.5
supervisor==4.2.0
scandir==1.9.0
selenium==3.141.0
sentry-sdk==0.10.2
shellescape==3.4.1
simplejson==3.10.0
simplekv==0.12.0
six==1.11.0
speaklater==1.3
-e git+https://github.com/Nibblecomm/spotipo-plugins.git@96247f72a56bd0d48dc89eb6b4e1bb407c381fdd#egg=Spotipo_Plugins
SQLAlchemy==1.2.14
stripe==2.29.4
termcolor==1.1.0
terminaltables==3.1.0
text-unidecode==1.2
timezones==2.0
tweepy==3.7.0
twilio==5.7.0
typing-extensions==3.7.4.2
ua-parser==0.10.0
unicodecsv==0.14.1
-e git+https://github.com/unifispot/unifispot-datatables.git@0ab0a4b74370f9886bf8dd93315112e0fbb375e0#egg=unifispot_datatables
-e git+https://github.com/unifispot/unifispot-migrate.git@e99a170c7c49c2d02ec4540521fef4d443028221#egg=Unifispot_Migrate
uritemplate==3.0.0
urllib3==1.24.3
user-agents==2.1
uWSGI==2.0.17
validate-email==1.3
validators==0.11.2
venusian==1.2.0
vine==1.3.0
wcwidth==0.1.9
webassets==0.12.1
Werkzeug==1.0.1
WTForms==2.1
yarl==1.4.2
zipp==3.1.0